Extract ConcurrentLruCache for reuse in NamedParameterJdbcTemplate
Closes gh-24197
